refactor: split GUI surface modules
